profile, saved Properties and Searched Properties code changes
