+++ src/Shared/Components/index.ts
@@ -20,3 +20,6 @@ export { default as Properties } from './Properties/Properties'
 export { default as Messages } from './Messages/Messages'
 export { default as UnreadMessageCount } from './UnreadMessageCount/UnreadMessageCount'
 export { default as FeaturedProperties } from './FeaturedProperties/FeaturedProperties'
+export { default as SavedProperties } from './SavedProperties/SavedProperties'
+export { default as SearchResults } from './SearchResults/SearchResults'
+export { default as Profile } from './Profile/Profile'
