+++ src/app/properties/saved/page.tsx
@@ -1,64 +1,6 @@
-// "use client"
-// import { PropertyCard, PropertyModel, Spinner } from "@/Shared"
-// import { apiUrls } from "@/Shared/Tools"
-// import { useState, useEffect } from "react"
-// import { toast } from "react-toastify"
-
-// const SavedPropertiesPage = () => {
-//   const [properties, setProperties] = useState<PropertyModel[]>([])
-//   const [loading, setLoading] = useState(true)
-
-//   useEffect(() => {
-//     fetchSavedProperties()
-//   }, [])
-
-//   const fetchSavedProperties = async () => {
-//     try {
-//       const res = await fetch(apiUrls.bookmarks)
-
-//       if (res.status === 200) {
-//         const data = await res.json()
-//         setProperties(PropertyModel.deserializeList(data))
-//       } else {
-//         console.log(res.statusText)
-//         toast.error("Failed to fetch saved properties")
-//       }
-//     } catch (error) {
-//       console.log(error)
-//       toast.error("Failed to fetch saved properties")
-//     } finally {
-//       setLoading(false)
-//     }
-//   }
-
-//   return loading ? (
-//     <Spinner />
-//   ) : (
-//     <section className="px-4 py-6">
-//       <div className="container-xl lg:container m-auto px-4 py-6">
-//         <h1 className="text-2xl mb-4">Saved Properties</h1>
-//         {properties.length === 0 ? (
-//           <p>No saved properties</p>
-//         ) : (
-//           <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-//             {properties.map((property) => (
-//               <PropertyCard key={property.id} property={property} />
-//             ))}
-//           </div>
-//         )}
-//       </div>
-//     </section>
-//   )
-// }
-// export default SavedPropertiesPage
-
-
-import React from 'react'
+import { SavedProperties } from "@/Shared"
 
 const SavedPropertiesPage = () => {
-  return (
-    <div>SavedPropertiesPage</div>
-  )
+  return <SavedProperties />
 }
-
 export default SavedPropertiesPage
